+++ src/main/java/com/nemonotfound/nemoscarpentry/datagen/ModModels.java
@@ -19,6 +19,7 @@ public class ModModels {
     public static final Model CHAIR_LUKAS_NATURAL = block("seats/chair_lukas_natural", TextureKey.ALL, TextureKey.TEXTURE, TextureKey.BOTTOM, TextureKey.PARTICLE);
     public static final Model CHAIR_TOMMY_NATURAL = block("seats/chair_tommy_natural", TextureKey.ALL, TextureKey.TEXTURE, TextureKey.BOTTOM, TextureKey.PARTICLE);
     public static final Model TABLE_THE_CLASSIC = block("tables/table_the_classic", TextureKey.ALL, TextureKey.PARTICLE);
+    public static final Model STANDING_TABLE = block("tables/standing_table", TextureKey.ALL, TextureKey.TEXTURE, TextureKey.PARTICLE);
     public static final Model CHAIR = block("seats/chair", TextureKey.ALL, TextureKey.PARTICLE);
     public static final Model CHAIR_FELIX = block("seats/chair_felix", TextureKey.ALL, TextureKey.PARTICLE);
     public static final Model LADDER = block("ladders/ladder", TextureKey.TEXTURE, TextureKey.PARTICLE);